pyproject.toml

[project]
dependencies = ["PyGithub==1.59.1", "feedgen", "marko", "markdown", "feedparser", "pyyaml", "requests", "openai"]
requires-python = ">=3.13"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
